#102 New emails for assigned and finished learning
